+++ src/js/vdlx-datagrid/datagrid.js
@@ -21,7 +21,8 @@
     limitations under the License.
  */
 import Tabulator from 'tabulator-tables/dist/js/tabulator';
-import {insightModules, insight} from '../insight-globals';
+import { insightModules, insight } from '../insight-globals';
+import {createSorter, createFormattedSorter} from './datagrid-sorter';
 import dataTransform, {
     getAllColumnIndices,
     getDisplayIndices,
@@ -31,13 +32,13 @@ import dataTransform, {
 import withScenarioData from './data-loader';
 import exportCsv from './export-csv';
 import Paginator from './paginator';
-import {getRowData} from './utils';
-import {EDITOR_TYPES} from '../constants';
+import { getRowData } from './utils';
+import { EDITOR_TYPES } from '../constants';
 import AddRemove from './add-remove';
-import {chooseColumnFilter} from './grid-filters';
-import perf from '../performance-measurement';
-import {createStateManager} from './state-peristence';
-import {DatagridLock} from './datagrid-lock';
+import { chooseColumnFilter } from './grid-filters';
+import {perf, perfMessage} from '../performance-measurement';
+import { createStateManager } from './state-peristence';
+import { DatagridLock } from './datagrid-lock';
 import escape from 'lodash/escape';
 import delay from 'lodash/delay';
 import some from 'lodash/some';
@@ -60,11 +61,17 @@ import each from 'lodash/each';
 import noop from 'lodash/noop';
 import isEmpty from 'lodash/isEmpty';
 import isArray from 'lodash/isArray';
+import isObject from 'lodash/isObject';
 import sortBy from 'lodash/sortBy';
+import isEqual from 'lodash/isEqual';
+import cloneDeep from 'lodash/cloneDeep';
+import constant from "lodash/constant";
+import { withDeferred } from '../ko-utils';
 
 const SelectOptions = insightModules.load('components/autotable-select-options');
 const DataUtils = insightModules.load('utils/data-utils');
 const dialogs = insightModules.load('dialogs');
+const Enums = insightModules.load('enums');
 
 const SELECTION_CHANGED_EVENT = 'selection-changed';
 const SELECTION_REMOVED_EVENT = 'selection-removed';
@@ -72,7 +79,7 @@ const SELECTION_REMOVED_EVENT = 'selection-removed';
 const addSelectNull = items => {
     if (isArray(items)) {
         // add empty option to the start of the list
-        return [{key: undefined, value: ''}].concat(items);
+        return [{ key: undefined, value: '' }].concat(items);
     }
     return items;
 };
@@ -107,20 +114,22 @@ class Datagrid {
         this.columnOptions$ = columnOptions$;
         this.componentRoot = root;
         this.view = insight.getView();
-        this.schema = this.view.getProject().getModelSchema();
+        this.schema = this.view.getApp().getModelSchema();
 
         const options = ko.unwrap(gridOptions$);
 
         this.table = this.createTable(options);
+        this.tableLock = new DatagridLock(this.table.element);
+
+        let rowIndex = 1;
+        this.getRowIndex = () => rowIndex++;
 
         this.headerToolbar = root.querySelector('.header-toolbar');
         const footerToolbar = root.querySelector('.footer-toolbar');
 
-        this.addRemoveRowControl = this.createAddRemoveRowControl(footerToolbar, this.table, options);
+        this.addRemoveRowControl = this.createAddRemoveRowControl(footerToolbar, options);
         this.paginatorControl = this.createPaginatorControl(footerToolbar, this.table, options);
         this.stateManager = null;
-
-        this.tableLock = new DatagridLock(this.table.element);
 
         this.buildTable();
         this.update();
@@ -146,38 +155,51 @@ class Datagrid {
         this.savingPromise = Promise.resolve();
 
         this.viewUnloadHandler = () => {
-            return Promise.resolve(this.savingPromise).catch((err) => dialogs.toast(err.message, dialogs.level.ERROR));
+            return Promise.resolve(this.savingPromise).catch(err => dialogs.toast(err.message, dialogs.level.ERROR));
         };
 
         this.unloadHandlerId = this.view.addUnloadHandler(this.viewUnloadHandler);
     }
 
-
     buildTable() {
         const columnOptions$ = this.columnOptions$;
         const gridOptions$ = this.gridOptions$;
-        const {data: scenariosData$, errors: errors$} = withScenarioData(columnOptions$);
+        const { data: scenariosData$, errors: errors$ } = withScenarioData(columnOptions$);
 
-        this.subscriptions = this.subscriptions.concat(
+        const allOptions$ = withDeferred(
+            ko.pureComputed(() => {
+                if (!gridOptions$() || !columnOptions$() || !scenariosData$()) {
+                    return undefined;
+                }
+                return {
+                    gridOptions: gridOptions$(),
+                    columnOptions: columnOptions$(),
+                    scenariosData: scenariosData$()
+                };
+            })
+        );
+
+        this.subscriptions = this.subscriptions.concat([
             ko
                 .pureComputed(() => {
-                    const gridOptions = ko.unwrap(gridOptions$());
-                    const columnOptions = columnOptions$();
-                    const scenariosData = scenariosData$();
-                    const errors = errors$();
-
-                    if (errors) {
+                    if (errors$()) {
                         this.componentRoot.style.display = 'none';
                     } else {
                         this.componentRoot.style.display = 'block';
                     }
+                })
+                .subscribe(noop),
+            withDeferred(
+                ko.pureComputed(() => {
+                    const allOptions = allOptions$();
+                    if (allOptions) {
+                        const { gridOptions, columnOptions, scenariosData } = allOptions;
 
-                    if (!isEmpty(get(columnOptions, 'columnOptions'))) {
-                        this.tableLock.lock();
-                    }
+                        if (!isEmpty(get(columnOptions, 'columnOptions'))) {
+                            this.tableLock.lock();
+                        }
 
-                    if (gridOptions && columnOptions && scenariosData) {
-                        return perf('PERF TOTAL:', () =>
+                        return perf('vdlx-datagrid total build time:', () =>
                             this.setColumnsAndData(gridOptions, columnOptions, scenariosData).then(() =>
                                 this.tableLock.unlock()
                             )
@@ -185,8 +207,8 @@ class Datagrid {
                     }
                     return undefined;
                 })
-                .subscribe(noop)
-        );
+            ).subscribe(noop)
+        ]);
     }
 
     update() {
@@ -199,9 +221,13 @@ class Datagrid {
 
     saveState() {
         if (this.stateManager) {
+            let sorters = map(this.table.getSorters(), sorter => ({ dir: sorter.dir, column: sorter.field }));
+            if (isEqual(this.initialSortOrder, sorters)) {
+                sorters = [];
+            }
             const state = {
                 filters: this.table.getHeaderFilters(),
-                sorters: map(this.table.getSorters(), sorter => ({dir: sorter.dir, column: sorter.field}))
+                sorters: sorters
             };
 
             this.stateManager.saveState(state);
@@ -241,6 +267,9 @@ class Datagrid {
         };
         const saveState = () => this.saveState();
 
+        let sortPromise = Promise.resolve();
+        let sortPromiseResolve = noop;
+
         const tabulatorOptions = {
             pagination: options.pagination,
             paginationSize: options.paginationSize,
@@ -252,7 +281,18 @@ class Datagrid {
             height: '100%',
             resizableColumns: false,
             dataFiltered: saveState,
-            dataSorting: saveState,
+            dataSorting: () => {
+                this.tableLock && this.tableLock.lock();
+                sortPromise = new Promise(resolve => {
+                    sortPromiseResolve = resolve;
+                });
+                perf('datagrid sorting', constant(sortPromise));
+                saveState();
+            },
+            dataSorted: () => {
+                sortPromiseResolve();
+                this.tableLock && this.tableLock.unlock();
+            },
             cellEditing: cell => select(cell.getRow()),
             rowClick: (e, row) => select(row),
             rowSelectionChanged: (data, rows) => this.setSelectedRow(first(rows)),
@@ -297,8 +337,9 @@ class Datagrid {
         const columnsWidth = this.table.columnManager.getWidth();
 
         if (columnsWidth < tableWidth || inRange(columnsWidth, tableOffsetWidth - 2, tableOffsetWidth + 2)) {
-            const columns = filter(reject(this.table.getColumns(), column => !!column.getDefinition().width), column =>
-                column.getVisibility()
+            const columns = filter(
+                reject(this.table.getColumns(), column => !!column.getDefinition().width),
+                column => column.getVisibility()
             );
             const toAddPx = (tableWidth - columnsWidth) / columns.length;
 
@@ -323,7 +364,7 @@ class Datagrid {
                 rowData: rowData,
                 value: cell.getValue(),
                 element: cell.getElement(),
-                displayPosition: {row: rowPosition, column: cellIndex}
+                displayPosition: { row: rowPosition, column: cellIndex }
             });
 
             const cells = map(row.getCells(), getCell);
@@ -380,17 +421,16 @@ class Datagrid {
 
     /**
      * @param {Element} footerToolbar
-     * @param {*} table
      * @param {*} options
      * @returns
      * @memberof Datagrid
      */
-    createAddRemoveRowControl(footerToolbar, table, options) {
+    createAddRemoveRowControl(footerToolbar, options) {
         if (!options.addRemoveRow) {
             return undefined;
         }
 
-        const addRemoveControl = new AddRemove(table, options.addRemoveRow === 'addrow-autoinc');
+        const addRemoveControl = new AddRemove(this.table, this.getRowIndex, options.addRemoveRow === 'addrow-autoinc');
         addRemoveControl.setEnabled(false);
         addRemoveControl.appendTo(footerToolbar);
 
@@ -420,6 +460,7 @@ class Datagrid {
         const schema = this.schema;
         const indicesOptions = columnOptions.indicesOptions;
         const entitiesOptions = columnOptions.columnOptions;
+        const calculatedColumnsOptions = columnOptions.calculatedColumnsOptions;
         const allColumnIndices = getAllColumnIndices(schema, entitiesOptions);
 
         const setNameAndPosns = getDisplayIndices(allColumnIndices, entitiesOptions);
@@ -433,8 +474,10 @@ class Datagrid {
 
         const allScenarios = uniq([scenariosData.defaultScenario].concat(values(scenariosData.scenarios)));
 
+        const tabulatorSorters = this.table.modules.sort.sorters;
+
         const indicesColumns = map(setNamePosnsAndOptions, setNameAndPosn => {
-            const {name, options} = setNameAndPosn;
+            const { name, options } = setNameAndPosn;
             const entity = schema.getEntity(name);
             const displayEntity = resolveDisplayEntity(schema, entity);
             const isNumberEntity = DataUtils.entityTypeIsNumber(displayEntity);
@@ -447,11 +490,25 @@ class Datagrid {
                 }
                 return classes.join(' ');
             };
+
+            const defaultFormatter = cell => SelectOptions.getLabel(schema, allScenarios, entity, cell.getValue());
+
+            const getFormatter = (type = 'display') => {
+                if (options.render) {
+                    return cell =>
+                        options.render(cell.getValue(), type, getRowDataForColumns(cell.getData()));
+                }
+                return defaultFormatter;
+            };
+
             let column = assign({}, setNameAndPosn.options, {
                 title: escape(String(title)),
                 field: options.id,
                 cssClass: getClass(),
-                formatter: cell => SelectOptions.getLabel(schema, allScenarios, entity, cell.getValue()),
+                formatter: getFormatter(),
+                sorter: options.sortByFormatted
+                    ? createFormattedSorter(options.id, getFormatter('sort'), tabulatorSorters)
+                    : createSorter(displayEntity, tabulatorSorters),
                 dataType: entity.getType(),
                 elementType: displayEntity.getElementType(),
                 labelsEntity: entity.getLabelsEntity(),
@@ -501,20 +558,12 @@ class Datagrid {
                     .removeFromArray(entityOptions.name, rowKey)
                     .commit();
 
-            const getRowKey = flowRight(
-                rowData => {
-                    const tableKeys = getPartialExposedKey(setNameAndPosns, rowData);
-                    return generateCompositeKey(
-                        tableKeys,
-                        setNameAndPosns,
-                        allColumnIndices[columnNumber],
-                        entityOptions
-                    );
-                },
-                getRowDataForColumns
-            );
+            const getRowKey = flowRight(rowData => {
+                const tableKeys = getPartialExposedKey(setNameAndPosns, rowData);
+                return generateCompositeKey(tableKeys, setNameAndPosns, allColumnIndices[columnNumber], entityOptions);
+            }, getRowDataForColumns);
 
-            const saveValue = (rowData, value) => setArrayElement({key: getRowKey(rowData), value: value});
+            const saveValue = (rowData, value) => setArrayElement({ key: getRowKey(rowData), value: value });
             const removeValue = rowData => removeArrayElement(getRowKey(rowData));
 
             const checkboxFormatter = cell => {
@@ -526,17 +575,17 @@ class Datagrid {
 
             const defaultFormatter = cell => SelectOptions.getLabel(schema, allScenarios, entity, cell.getValue());
 
-            const getFormatter = () => {
+            const getFormatter = (type = 'display') => {
                 if (entityOptions.render) {
                     return cell =>
-                        entityOptions.render(cell.getValue(), 'display', getRowDataForColumns(cell.getData()));
+                        entityOptions.render(cell.getValue(), type, getRowDataForColumns(cell.getData()));
                 }
 
-                if (entityOptions.editorType === EDITOR_TYPES.checkbox) {
+                if (entityOptions.editorType === EDITOR_TYPES.checkbox && type === 'display') {
                     return checkboxFormatter;
-                } else {
-                    return defaultFormatter;
                 }
+
+                return defaultFormatter;
             };
 
             const checkboxCellClickHandler = (e, cell) => {
@@ -572,11 +621,7 @@ class Datagrid {
                     } else if (entityOptions.editorOptions) {
                         getOptions = flow(
                             entityOptions.editorOptions,
-                            options =>
-                                SelectOptions.generateSelectOptionsFromValues(
-                                    options,
-                                    isNumberEntity
-                                ),
+                            options => SelectOptions.generateSelectOptionsFromValues(options, isNumberEntity),
                             entityOptions.selectNull ? addSelectNull : identity
                         );
                     }
@@ -640,7 +685,7 @@ class Datagrid {
                 return undefined;
             };
 
-            const cellEdited = (cell) => {
+            const cellEdited = cell => {
                 $(cell.getElement()).off('keyup');
                 const oldValue = isUndefined(cell.getOldValue()) ? '' : cell.getOldValue();
                 const value = cell.getValue();
@@ -652,7 +697,7 @@ class Datagrid {
                     dialogs.alert(validationResult.errorMessage, VALIDATION_ERROR_TITLE, () => {
                         defer(() => cell.edit(true));
                     });
-                    this.savingPromise = Promise.reject({message: validationResult.errorMessage});
+                    this.savingPromise = Promise.reject({ message: validationResult.errorMessage });
                 } else {
                     if (value !== oldValue) {
                         if (isUndefined(value) || value === '') {
@@ -672,7 +717,7 @@ class Datagrid {
                 }
             };
 
-            const cellEditCancelled = (cell) => {
+            const cellEditCancelled = cell => {
                 $(cell.getElement()).off('keyup');
                 const value = cell.getValue();
                 const validationResult = validateAndStyle(cell, value);
@@ -695,6 +740,10 @@ class Datagrid {
                 cssClass: getClasses(),
                 cellClick: getCellClickHandler(),
                 formatter: getFormatter(),
+                sortByFormatted: entityOptions.sortByFormatted,
+                sorter: entityOptions.sortByFormatted
+                    ? createFormattedSorter(entityOptions.id, getFormatter('sort'), tabulatorSorters)
+                    : createSorter(displayEntity, tabulatorSorters),
                 editor: entityOptions.editorType,
                 editorParams: getEditorParams(),
                 cellEditing: getCellEditingHandler(),
@@ -720,9 +769,9 @@ class Datagrid {
                         const uncheckedValue = get(entityOptions, 'uncheckedValue', false);
                         return {
                             values: [
-                                {value: undefined, label: 'No Filter'},
-                                {value: String(checkedValue), label: 'Checked'},
-                                {value: String(uncheckedValue), label: 'Unchecked'}
+                                { value: undefined, label: 'No Filter' },
+                                { value: String(checkedValue), label: 'Checked' },
+                                { value: String(uncheckedValue), label: 'Unchecked' }
                             ]
                         };
                     }
@@ -802,11 +851,49 @@ class Datagrid {
             return column;
         });
 
-        let columns = sortBy([].concat(indicesColumns, entitiesColumns), (column) => column.index || -1)
+        const calculatedColumns = map(calculatedColumnsOptions, options => {
+            const title = get(options, 'title', options.name);
+
+            const getFormatter = (type = 'display') => cell => options.render(cell.getValue(), type, getRowDataForColumns(cell.getData()));
+
+            let column = assign({}, options, {
+                title: escape(String(title)),
+                formatter: getFormatter(),
+                name: options.name,
+                field: options.id,
+                elementType: Enums.DataType.STRING,
+                sortByFormatted: true,
+                sorter: createFormattedSorter(options.id, getFormatter('sort'), tabulatorSorters),
+                accessorDownload: (value, rowData) => options.render(value, 'display', getRowDataForColumns(rowData)),
+            });
+
+            if (gridOptions.columnFilter) {
+                const getHeaderFilterFn = () => {
+                    const columnFilter = chooseColumnFilter(column);
+                    if (columnFilter) {
+                        return (valueTxt, cellValue, rowData, params) => {
+                            var value = column.render(cellValue, 'filter', getRowDataForColumns(rowData));
+                            return columnFilter(valueTxt, value, rowData, params);
+                        };
+                    }
+                    return undefined;
+                };
+
+                column = assign(column, {
+                    headerFilterPlaceholder: 'No filter',
+                    headerFilter: !!gridOptions.columnFilter,
+                    headerFilterFunc: getHeaderFilterFn()
+                });
+            }
+
+            return column;
+        });
+
+        let columns = sortBy([].concat(indicesColumns, entitiesColumns, calculatedColumns), column => column.index || -1);
 
         let freezeColumns = parseInt(gridOptions.freezeColumns);
         if (freezeColumns && !isNaN(freezeColumns)) {
-            columns = map(columns, function (col, idx) {
+            columns = map(columns, function(col, idx) {
                 if (idx < freezeColumns) {
                     col.frozen = true;
                 }
@@ -814,18 +901,22 @@ class Datagrid {
             });
         }
 
-        const {data, allSetValues} = perf('PERF Data generation:', () =>
+        const { data, allSetValues } = perf('Data generation:', () =>
             dataTransform(
                 allColumnIndices,
                 columns,
                 entitiesColumns,
                 setNamePosnsAndOptions,
                 scenariosData,
-                gridOptions.rowFilter
+                gridOptions.rowFilter,
+                this.getRowIndex
             )
         );
 
-        const editable = some(reject(entitiesOptions, options => !get(options, 'visible', true)), 'editable');
+        const editable = some(
+            reject(entitiesOptions, options => !get(options, 'visible', true)),
+            'editable'
+        );
         if (!editable && gridOptions.addRemoveRow) {
             console.log(
                 `vdl-table (${gridOptions.tableId}): add/remove rows disabled. Table needs to have at least one editable column to use this feature.`
@@ -844,12 +935,27 @@ class Datagrid {
         this.indicesColumns = indicesColumns;
 
         table.setColumns(columns);
+        this.initialSortOrder = map(
+            sortBy(
+                filter(columns, column => !isUndefined(column.sortOrder)),
+                'sortOrder'
+            ),
+            column => ({
+                column: column.id,
+                dir: column.sortDirection
+            })
+        );
+
+        this.table.setSort(cloneDeep(this.initialSortOrder));
 
         this.stateManager = this.createStateManager(gridOptions, columns, map(entitiesColumns, 'scenario'));
         this.loadState();
 
         const redraw = () => {
-            if (this.table.element.offsetParent) {
+            if (
+                this.table.element.offsetParent &&
+                this.table.element.offsetParent.tagName.toLowerCase() === 'vdlx-datagrid'
+            ) {
                 return Promise.resolve(this.table.redraw(true));
             } else {
                 return new Promise((resolve, reject) => {
@@ -861,12 +967,25 @@ class Datagrid {
                 });
             }
         };
-        return perf('PERF Tabulator.setData():', () =>
+
+        this.table.element.style.visibility = 'hidden';
+
+        perfMessage(() => {
+            if (isArray(data) && isObject(data[0])) {
+                let columnCount = Object.keys(data[0]).length - 1;
+                let rowCount = data.length;
+                let cellCount = rowCount * columnCount;
+                return `vdlx-datagrid going to render with ${rowCount.toLocaleString()} rows, ${columnCount} columns, ${cellCount.toLocaleString()} cells`;
+            }
+        });
+
+        return perf('Tabulator setData and draw', () =>
             table
                 .setData(data)
                 .then(() => redraw())
-                .catch(err => {
-                    debugger;
+                .then(() => (this.table.element.style.visibility = 'visible'))
+                .catch(e => {
+                    console.error('An error occurred whilst adding data to Tabulator and redrawing', e);
                 })
         );
     }